abstract addProduct in an own function to share it through module system
