+++ src/components/List/CardList.tsx
@@ -6,13 +6,12 @@ import style from './List.module.css';
 import IDataJson from '../DataJson';
 import { Link } from 'react-router-dom';
 import IProductShoppingCart from '../IProductShoppingCart';
+import addProduct from '../../utils/Products';
 
 const CardList = ({
-    data,
-    addP
+    data
   }: {
     data: Array<IDataJson>;
-    addP: (product: IProductShoppingCart) => void;
   }) : JSX.Element => {
   return (
     <GridList cols={4} cellHeight={"auto"} spacing={1} className={style.gridList}>
@@ -23,7 +22,7 @@ const CardList = ({
               pathname: `/products/${shirt.id}`,
               state: {data: shirt}
             }} >
-              <Card id={shirt.id} dark={true} img={shirt.img} title={shirt.title} price={shirt.price} addP={addP}/>
+              <Card id={shirt.id} dark={true} img={shirt.img} title={shirt.title} price={shirt.price} addP={addProduct}/>
             </Link>
           </div>
