+++ src/components/List/ImageList.tsx
@@ -5,19 +5,17 @@ import GridListTile from '@material-ui/core/GridListTile';
 import GridListTileBar from '@material-ui/core/GridListTileBar';
 import IconButton from '@material-ui/core/IconButton';
 import IDataJson from '../DataJson';
-import IProductShoppingCart from '../IProductShoppingCart';
 import style from './List.module.css';
 import { Link } from 'react-router-dom';
+import addProduct from '../../utils/Products';
 
 const ImageList  = ({
-  data,
-  addP
+  data
 }: {
   data: Array<IDataJson>;
-  addP: (product: IProductShoppingCart) => void;
 }) : JSX.Element => {
 
-const addProduct = (e) => {
+const addProductAux = (e) => {
   console.log(e.currentTarget);
   const shirt = JSON.parse(e.currentTarget.value);
   const product = {
@@ -27,7 +25,7 @@ const addProduct = (e) => {
     price: shirt.price,
     qty: 1
   };
-  addP(product);
+  addProduct(product);
 };
 
 
@@ -52,7 +50,7 @@ return (
                 color={"primary"} 
                 className={style.icon} 
                 value={JSON.stringify(shirt)}
-                onClick={addProduct}>
+                onClick={addProductAux}>
                   <AddShoppingCartIcon />
               </IconButton>
             }
