+++ src/components/Product/Product.tsx
@@ -4,7 +4,7 @@ import SelectButton from '../SelectButton';
 import IconButton from '@material-ui/core/IconButton';
 import AddShoppingCartIcon from '@material-ui/icons/AddShoppingCart';
 import style from './Product.module.css';
-
+import addProduct from '../../utils/Products';
 
   const Product = ({data}:{data:IDataJson}): JSX.Element =>  {
     const [state, setState] = React.useState<{ qty: number}>({
@@ -18,9 +18,8 @@ import style from './Product.module.css';
       });
     };
 
-    const addProduct = (event: React.ChangeEvent) => {
+    const addProductAux = (event: React.ChangeEvent) => {
       event.preventDefault();
-      const old = JSON.parse(localStorage.getItem('shoppingcart'));
       const aux={
         id: data.id,
         img: data.img,
@@ -28,26 +27,7 @@ import style from './Product.module.css';
         price: data.price * state.qty,
         qty: state.qty
       };
-      if(old === null){
-        const array=[];
-        array.push(aux);
-        localStorage.setItem('shoppingcart', JSON.stringify(array));
-      }
-      else{
-        const element= old.find(p => p.id === aux.id);
-        if(element){
-          const newList= old.filter(p=> p.id !== aux.id);
-          aux.qty= aux.qty + parseInt(element.qty,10);
-          aux.price= aux.price + parseInt(element.price,10);
-          newList.push(aux);
-          localStorage.setItem('shoppingcart', JSON.stringify(newList));
-        }
-        else{
-          old.push(aux);
-          localStorage.setItem('shoppingcart', JSON.stringify(old));
-        }
-      }
-      alert('product added');
+      addProduct(aux);
     }
 
     const values=[];
@@ -66,7 +46,7 @@ import style from './Product.module.css';
           <div className={style.centeredRight}>
               <h2 className={style.description}>{data.description}</h2>
               <div className={style.containerButton}>
-                <IconButton   aria-label={`star ${data.title}`} color={"default"}  onClick={addProduct} >
+                <IconButton   aria-label={`star ${data.title}`} color={"default"}  onClick={addProductAux} >
                       <AddShoppingCartIcon />
                       <p>Add</p>
                 </IconButton>
